requirements.txt
aiofiles==24.1.0
aiohttp==3.11.14
beautifulsoup4==4.13.3
Flask==3.1.0
flask_cors==5.0.1
mysql_connector_repackaged==0.3.1
pandas==2.2.3
pdfplumber==0.11.5
tabulate==0.9.0
